extract test_synonyms_list
