+++ solr-admin-app/tests/conftest.py
@@ -11,7 +11,7 @@ def port():
 @pytest.fixture(scope="session")
 def server(port):
     server = ServerDriver(name='MyServer', port=port)
-    server.start(cmd=['gunicorn', 'app', '-w', '1', '-b', '0.0.0.0:{0}'.format(port)])
+    server.start(cmd=['python', 'app.py'])
     return server
 
 
@@ -41,17 +41,13 @@ def base_url(port, server):
 
 
 @pytest.fixture(scope="session")
-def schema():
-    return open('../../database/database_create.sql').read()
-
-
-@pytest.fixture(scope="session")
-def db(schema):
+def db():
     from flask_sqlalchemy import SQLAlchemy
     from solr_admin import create_application
     app = create_application(run_mode='testing')
     db = SQLAlchemy(app)
 
+    schema = open('database/database_create.sql').read()
     sqls = [sql for sql in [x.replace('\n', '').strip() for x in schema.split(';')] if len(sql) > 0]
 
     for sql in sqls:
